fix: pass only _client/_aclient properties to context SPARQLWrapper
The idea of SPARQLWrapper context manager is to provide an interface
for automatic client management that still enables re-use of a client.
For that, the context manager exposes a fresh SPARQLWrapper instance
that gets initialized with clients that would otherwise be used by
SPARQLWrapper internally (either a newly created client or a user
supplied client).

That means that the context client needs to be instantiated with
both the _client/_aclient properties and does not need
client_config/aclient_config.