refactor: implement ClientManager component for client sharing logic
The change separates out the client sharing logic into a ClientManager
component. This allows for a much cleaner and much less cluttered
design of the sparqlx.SPARQLWrapper class.

Note that the ClientManager does not implement the
ContextManager/AsyncContextmanager protocol itself but exposes
the contextlib.contextmanagers context/acontext. This design is
itentional and leverages the fact, that a contextlib.contextmanager
can reference the same object between enter and
exit without managing a context stack (or the like).
Implementing the equivalent logic with __enter__/__exit__ is
possible but has significantly more overhead, especially for nested
context managers which would need to keep track of entered/exited contexts.