feat: add SPARQLParseException for signalling parsing failure
The change adds a SPARQLParseException exception that gets raised on
any failure resulting from calls to RDFLib prepareQuery.

This allows client code to check for a sparqlx-controlled exception and avoids
using the transitive pyparsing.ParseExceptions raised by RDFLib both for
sparqlx and library users.

Note that parsing and "preparing" (translating the parse tree to
SPARQL algebra) is differentiated in the RDFLib SPARQL parser;
catching any exception from prepareQuery and re-raising it as
sparqlx.SPARQLParseException is still semantically consistent for a
broader notion of parsing.

Closes #82.