feat: implement SPARQL Protocol Operation types
The change implements support for all SPARQL Protocol HTTP Operation
types according to the SPARQL 1.2 Protocol Specification.

The spec defines

- query via GET
- query via URL-encoded POST
- query via POST directly

for Query Operations and

- query via URL-encoded POST
- query via POST directly

for Update Operations.

The commit introduces a `query_method: Literal["POST",
"POST-direct"]` and a `update_method: Literal["POST", "POST-direct""]`
parameter to the `SPARQLWrapper` class that allows to set the SPARQL
Operation type.

The "query via URL-encoded POST" option is set as default
for both query and update operations as it allows to send a Request
Content Type in the Accept Header and both the Query/Update Request
strings and Query/Update Parameters in the Request Message Body, which
is arguably the most convenient HTTP operation option defined in the spec.